clean code and update reader.py
