--- QA/reader.py
@@ -1,35 +1,0 @@
-from transformers import DistilBertTokenizer, DistilBertForQuestionAnswering
-import torch
-
-tokenizer = DistilBertTokenizer.from_pretrained(
-    'distilbert-base-uncased-distilled-squad')
-model = DistilBertForQuestionAnswering.from_pretrained(
-    'distilbert-base-uncased-distilled-squad')
-
-
-def answer(question, text):
-    input_ids = torch.tensor(tokenizer.encode(
-        question, text, max_length=512, add_special_tokens=True)).unsqueeze(0)  # Batch size 1
-    start_positions = torch.tensor([1])
-    end_positions = torch.tensor([3])
-
-    outputs = model(input_ids, start_positions=start_positions,
-                    end_positions=end_positions)
-    loss, start_scores, end_scores = outputs[:3]
-
-    all_tokens = tokenizer.convert_ids_to_tokens(input_ids[0])
-    answer = tokenizer.convert_tokens_to_string(
-        all_tokens[torch.argmax(start_scores): torch.argmax(end_scores) + 1])
-    print('*************')
-    # print(answer)
-    return answer
-
-
-#######################################################################
-
-# corpus = "Lionel Andrés Messi Cuccittini (born 24 June 1987) is an Argentine professional footballer who plays as a forward and captains both Spanish club Barcelona and the Argentina national team. Often considered the best player in the world and widely regarded as one of the greatest players of all time, Messi has won a record six Ballon d'Or awards,[note 2] and a record six European Golden Shoes. He has spent his entire professional career with Barcelona, where he has won a club-record 34 trophies, including ten La Liga titles, four UEFA Champions League titles and six Copas del Rey. A prolific goalscorer and a creative playmaker, Messi holds the records for most goals in La Liga (438), a La Liga and European league season (50), most hat-tricks in La Liga (36) and the UEFA Champions League (8), and most assists in La Liga (181) and the Copa América (12). He has scored over 700 senior career goals for club and country.Born and raised in central Argentina, Messi relocated to Spain to join Barcelona at age 13, for whom he made his competitive debut aged 17 in October 2004. He established himself as an integral player for the club within the next three years, and in his first uninterrupted season in 2008–09 he helped Barcelona achieve the first treble in Spanish football; that year, aged 22, Messi won his first Ballon d'Or. Three successful seasons followed, with Messi winning three consecutive Ballons d'Or, making him the first player to win the award four times. During the 2011–12 season, he set the La Liga and European records for most goals scored in a single season, while establishing himself as Barcelona's all-time top scorer. The following two seasons, Messi finished second for the Ballon d'Or behind Cristiano Ronaldo (his perceived career rival), before regaining his best form during the 2014–15 campaign, becoming the all-time top scorer in La Liga and leading Barcelona to a historic second treble, after which he was awarded a fifth Ballon d'Or in 2015. Messi assumed the captaincy of Barcelona in 2018, and in 2019 he secured a record sixth Ballon d'Or.An Argentine international, Messi is his country's all-time leading goalscorer. At youth level, he won the 2005 FIFA World Youth Championship, finishing the tournament with both the Golden Ball and Golden Shoe, and an Olympic gold medal at the 2008 Summer Olympics. His style of play as a diminutive, left-footed dribbler drew comparisons with his compatriot Diego Maradona, who described Messi as his successor. After his senior debut in August 2005, Messi became the youngest Argentine to play and score in a FIFA World Cup during the 2006 edition, and reached the final of the 2007 Copa América, where he was named young player of the tournament. As the squad's captain from August 2011, he led Argentina to three consecutive finals: the 2014 FIFA World Cup, for which he won the Golden Ball, and the 2015 and 2016 Copas América. After announcing his international retirement in 2016, he reversed his decision and led his country to qualification for the 2018 FIFA World Cup, and a third-place finish at the 2019 Copa América.One of the most famous athletes in the world, Messi has been sponsored by sportswear company Adidas since 2006 and has established himself as their leading brand endorser. According to France Football, he was the world's highest-paid footballer for five years out of six between 2009 and 2014, and was ranked the world's highest-paid athlete by Forbes in 2019. Messi was among Time's 100 most influential people in the world in 2011 and 2012. In 2020, he was awarded the Laureus World Sportsman of the Year. "
-text = "Looking to kickstart your career with several months of intensive immersion at a major banking group? That is the opportunity offered to recent graduates by BNP Paribas’ Graduate Programs. These challenging 18-month programs spent in different operational functions enable participants to expand their banking knowledge, skills and professional networks. Learn more about this promising initiative. Convinced that recent and future graduates represent the future of the Group, BNP Paribas is counting on their talent to maintain its edge in the market. As it works to build the future of banking, the Group is now putting in place HR measures designed to attract future talent. The Graduate Programs are one example: these recruiting programs allow selected candidates to join several operational functions for 18 months, while benefiting from personalized HR support. Hired immediately through long-term contracts, participants in each pathway complete an immersive and personalized curriculum composed of three professional development assignments, each lasting six months. The rotation is co-constructed based on the needs of the bank’s businesses and the skills or interests of each participant. As full team members, participants quickly gain experience and specific skills by working directly with their peers and senior employees in the entity. As soon as they are hired, participants also become members of a “trainee class” to promote joint development, experience-sharing and to help build skills as a group. Recruiting programs are open to candidates of all nationalities, though they take place in France (with some exceptions). The programs seek a wide range of candidates. While Master’s (Bac+5) graduates in math, finance, economics, science, business, engineering and computer science, showing strong analytical skills and mathematical abilities, can apply for the Excellence Program, digital and data-oriented candidates are invited to join the Digital Path. The Digital Path encourages a passionate mindset and a “digital explorer” mentality, rather than a specific degree. In this way, the Group aims to develop an internal mindset focused on digital innovation, notably by integrating the following candidates: For all programs, initiative, determination and curiosity are essential qualities. Participants should also be committed team players eager to tackle collaborative work. Candidates need not have a specific end goal in mind for the program—based on their preferences and the opportunities available with each business, participants can co-construct their pathway with HR teams as they progress through the program.BNP Paribas has long maintained a strong relationship with the academic world and target schools in order to attract its future talent, whether it be universities, business schools or engineering schools. The Group serves as a committed partner of the academic world. It also plays a role in curriculum planning by updating coursework so that it matches actual business needs as closely as possible. To promote the bank’s businesses among students and recruit high-potential candidates or future talent, the BNP Paribas Group relies on its Campus Management team, which ensures that the bank’s initiatives carried out with certain schools in France match the needs of its businesses, departments and subsidiaries. In other words, the role of Campus Management is to develop quality relationships with students and promote the Group’s businesses. In the words of Jean-Dominique Criscuolo, Manager of Partnerships and Relations with Schools and Universities, “We strongly believe that our Academic Relations play a growing role in transforming our Group.” According to Jean-Dominique Criscuolo, BNP Paribas comprises some 300 businesses, some of which only emerged within the last two years. Including data scientists, agile coaches, and IT inspectors, the new professions created by the digital transformation broaden the palette of the Group’s traditional businesses. In a changing world, governed by rapidly evolving international financial regulations, many new opportunities are available within our Compliance teams. This is a dynamic profession which, due to regulatory, geopolitical and societal changes, plays an increasingly central role in the crossroads of strategy and the daily actions of the bank and its customers.” This large-scale project will further expand in coming years, notably internationally. Gilles Deschanel, in charge of academic partnerships at BNP Paribas CIB, observes that “this partnership aims to become one of the leading research bodies in this immensely disruptive technology. It will also play a role in transforming the financial sector through publications and major events like VivaTech.” Since January 2016, BNP Paribas has offered an Excellence Program targeting new Master’s level graduates (BAC+5) who show high potential. The aid program lasts 18 months and comprises three assignments of six months each. It serves as a strong career accelerator that enables participants to access high-level management positions at a faster rate. The program allows participants to discover the BNP Paribas Group and its various entities in France and abroad, build an internal and external network by working on different assignments and receive personalized assistance from a mentor and coaching firm at every step along the way."
-query = 'Since when does the Excellence Program of BNP Paribas exist?'
-#question = "What is the occupation of Messi ?"
-
-answer(query, text[:500])
